feature-gate-interface: Add revoke ix, reorganize
#### Problem

In order update Agave to SDK v3, it needs to use all crates that are
also on SDK v3. The Agave CLI uses solana-feature-gate-client, which is
auto-generated.

#### Summary of changes

Add what's required for the Agave CLI to just use
solana-feature-gate-interface, which is the revoke instruction and error
type.

While I was at it, I moved things around to make it look like a normal
interface crate, creating instruction.rs and state.rs.
